requirements.txt
# WayFare MVP Backend - Production Dependencies
# Generated for deployment requirements

# Core Framework Dependencies (from nanobot)
typer>=0.20.0,<1.0.0
litellm>=1.81.5,<2.0.0
pydantic>=2.12.0,<3.0.0
pydantic-settings>=2.12.0,<3.0.0
loguru>=0.7.3,<1.0.0
openai>=2.8.0

# Database
aiosqlite>=0.20.0,<1.0.0

# Document Parsing
PyMuPDF>=1.24.0,<2.0.0
markdown-it-py>=3.0.0,<4.0.0

# Embedding & Vector Store
onnxruntime>=1.20.0,<2.0.0
transformers>=4.50.0,<5.0.0
qdrant-client>=1.12.0,<2.0.0
numpy>=1.26.0,<2.0.0

# Hashing
blake3>=0.4.0,<1.0.0

# Data Validation & Config
pyyaml>=6.0.0,<7.0.0

# HTTP Client (for LLM API calls)
httpx>=0.28.0,<1.0.0

# Async Support
asyncio-compat>=0.1.0; python_version < "3.12"
